refactor : Connection 프록시 생성 로직을 별도의 클래스로 분리
